requirements.txt
CherryPy==8.1.0
click==6.6
decorator==4.0.10
dominate==2.2.1
Flask==0.11.1
Flask-Bootstrap==3.3.7.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Paste==2.0.3
pbr==1.10.0
RPi.GPIO==0.6.2
rrdtool==0.1.5
six==1.10.0
smbus2==0.1.4
SQLAlchemy==1.0.15
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
visitor==0.1.3
Werkzeug==0.11.11
WTForms==2.1
